backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
clickhouse-connect==0.7.0
python-multipart==0.0.6
pandas==2.1.3
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0 